Fix regressor cost
The regression loss function is wrong in tfvs. `tf.nn.l2_loss` returns a scalar (the sum of the per-example weights), which obviously will frequently get the weighted losses wrong. This PR calculates per-example losses, which are then weighted appropriately. (Sorry, this should have been fixed when tfvs was released.)