Propose update to Tensorboard integration
This PR is meant illustrate an update to tensorboard that I would like to propose, and a PR request seemed like a reasonable forum to do so. Currently, the add_summary_to_tg() called on each layer while building the graph limits the information transferred to Tensorboard only to the out tensor. Therefore on a Dense layer or GraphConvolution layer, for example, we can only see the how the activation changes over time and not necessarily the weights or biases, which in my experience can often provide more actionable information. This requires custom functions for each layer containing important variables, which is where I do not know how to proceed. The first pass solution I created is to add an add_summary_to_tg() function to the Dense and GraphConv layer which overwrites the parent function. The immediate issue with this change is that it does not rely on layer.add_summary(), and therefore the user cannot choose whether this information should be logged. This would require changing layer.add_summary(), which while not difficult, is something I wanted to consult with Deepchem team before doing. I have ideas about how to change this, but would like to discuss them before making large changes to the code. I look forward to hearing your feedback and thoughts, and let me know if this was the correct forum for this conversation.