Add files via upload
Introducing a tutorial for a new drug modality Antibody drug conjugates, We briefly go over the working of an ADC conceptually and we included code that implements a deep learning to predict ADC activity given the target protein sequence. It uses a simple Multi-Layered Perceptron to make the prediction which can be improved upon significantly later on.